+++ src/vldp/vldp_common.h
@@ -57,7 +57,7 @@ extern Uint32 g_req_frame;       // which frame to seek to
 extern Uint32 g_req_min_seek_ms; // minimum # of milliseconds that this seek can
                                  // take
 extern Uint32 g_req_timer;
-extern unsigned long g_req_idx; // multipurpose index
+extern uint32_t g_req_idx;       // multipurpose index
 extern VLDP_BOOL g_req_precache;
 extern char g_req_file[];        // which file to open
 extern Uint8 g_req_cmdORcount;   // the current command count OR'd with the
